+++ global_helpers/panther_base_helpers.py
@@ -184,3 +184,17 @@ def key_value_list_to_dict(list_objects: List[dict], key: str, value: str) -> di
     # example: [{'key': 'a', 'value': 1}, {'key': 'b', 'value': 2}]
     # becomes: {'a': 1, 'b': 2}
     return {item[key]: item[value] for item in list_objects}
+
+
+# When a single item is loaded from json, it is loaded as a single item
+# When a list of items is loaded from json, it is loaded as a list of that item
+# When we want to iterate over something that could be a single item or a list
+# of items we can use listify and just continue as if it's always a list
+def listify(maybe_list):
+    try:
+        iter(maybe_list)
+    except TypeError:
+        # not a list
+        return [maybe_list]
+    # either a list or string
+    return [maybe_list] if isinstance(maybe_list, (str, bytes, dict)) else maybe_list
