+++ packs/asana.yml
@@ -16,10 +16,8 @@ PackDefinition:
     - Asana.Workspace.New.Admin
     # Globals used in these detections
     - panther_asana_helpers
-    - panther_base_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
+
     - panther_event_type_helpers
     # Data Model
     - Standard.Asana.Audit
