+++ packs/atlassian.yml
@@ -5,10 +5,8 @@ PackDefinition:
   IDs:
     - Atlassian.User.LoggedInAsUser
     # Globals used in these detections
-    - panther_base_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
+
     - panther_event_type_helpers
     # Data Model
     - Standard.Atlassian.Audit
