+++ packs/auth0.yml
@@ -18,7 +18,5 @@ PackDefinition:
     - panther_base_helpers
     - panther_auth0_helpers
     - global_filter_auth0
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
 DisplayName: "Panther Auth0 Pack"
