+++ packs/aws.yml
@@ -190,12 +190,12 @@ PackDefinition:
     - Standard.OCSF.NetworkActivity
     - Standard.OCSF.DnsActivity
     # Globals used in these rules/policies
+    - panther_aws_helpers
     - panther_base_helpers
     - panther_config
     - panther_config_defaults
     - panther_config_overrides
-    - panther_aws_helpers
-    - panther_event_type_helpers
     - panther_iocs
+    - panther_ipinfo_helpers
     - panther_lookuptable_helpers
     - panther_oss_helpers
