+++ packs/azure_signin.yml
@@ -9,10 +9,8 @@ PackDefinition:
     # Globals used in these detections
     - global_filter_azuresignin
     - panther_azuresignin_helpers
-    - panther_base_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
+
     - panther_event_type_helpers
     # Data Models
     - Standard.Azure.Audit.SignIn
