+++ packs/box.yml
@@ -15,9 +15,6 @@ PackDefinition:
     # Globals used in these detections
     - panther_base_helpers
     - panther_box_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
     - panther_event_type_helpers
     # Data Models
     - Standard.Box.Event
