+++ packs/cloudflare.yml
@@ -11,9 +11,6 @@ PackDefinition:
     - panther_cloudflare_helpers
     - panther_lookuptable_helpers
     - global_filter_cloudflare
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
     # Data Models
     - Standard.Cloudflare.Firewall
     - Standard.Cloudflare.HttpReq
