+++ packs/credential_security.yml
@@ -25,6 +25,10 @@ PackDefinition:
     - panther_config
     - panther_config_defaults
     - panther_config_overrides
+    - panther_crowdstrike_fdr_helpers
+    - panther_msft_helpers
+    - panther_okta_helpers
+    - panther_slack_helpers
     # Rules
     - AWS.CloudTrail.RootPasswordChanged
     - AWS.IAM.AccessKeyCompromised
