+++ packs/crowdstrike.yml
@@ -20,9 +20,7 @@ PackDefinition:
     - Crowdstrike.Macos.Osascript.Administrator
     # Globals used in these detections
     - panther_base_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+    - panther_crowdstrike_fdr_helpers
     # Data models
     - Standard.AWS.VPCDns
     - Standard.CiscoUmbrella.DNS
