+++ packs/crowdstrike_event_streams.yml
@@ -5,10 +5,6 @@ PackDefinition:
   IDs:
     - panther_crowdstrike_event_streams_helpers
     - panther_base_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
-    
     - Crowdstrike.AdminRoleAssigned
     - Crowdstrike.AllowlistRemoved
     - Crowdstrike.API.Key.Created
