+++ packs/dropbox.yml
@@ -10,7 +10,7 @@ PackDefinition:
     - Dropbox.User.Disabled.2FA
     - Dropbox.Admin.sign.in.as.Session
     # Globals used in these detections
-    - panther_base_helpers
+
     - panther_config
     - panther_config_defaults
     - panther_config_overrides
