+++ packs/duo.yml
@@ -20,8 +20,6 @@ PackDefinition:
     - Duo.Admin.SSO.SAML.Requirement.Disabled
     - Duo.Admin.User.MFA.Bypass.Enabled
     # Globals used in these detections
-    - panther_base_helpers
+
     - panther_duo_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
