+++ packs/gcp_audit.yml
@@ -50,8 +50,5 @@ PackDefinition:
     # Globals used in these rules/policies
     - panther_gcp_helpers
     - panther_base_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
     - panther_event_type_helpers
 DisplayName: "Panther GCP Audit Pack"
