+++ packs/gcp_k8.yml
@@ -17,6 +17,3 @@ PackDefinition:
     # Globals
     - panther_gcp_helpers
     - panther_base_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
