+++ packs/github.yml
@@ -29,10 +29,7 @@ PackDefinition:
     # Data model
     - Standard.Github.Audit
     # Globals
-    - panther_base_helpers
+    - panther_github_helpers
     - panther_event_type_helpers
-    - panther_oss_helpers
     - global_filter_github
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
