+++ packs/gsuite_reports.yml
@@ -36,9 +36,7 @@ PackDefinition:
     - Standard.GSuite.Reports
     # Globals used in these detections
     - panther_base_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+    - panther_gsuite_helpers
     - panther_event_type_helpers
     - panther_lookuptable_helpers
     # Queries
