+++ packs/ipinfo.yml
@@ -12,7 +12,5 @@ PackDefinition:
     - panther_base_helpers
     - panther_ipinfo_helpers
     - panther_lookuptable_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
 DisplayName: "IPInfo"
