+++ packs/mongodb.yml
@@ -15,8 +15,6 @@ PackDefinition:
     - MongoDB.External.UserInvited.NoConfig
     - MongoDB.Logging.Toggled
     # Globals
-    - panther_base_helpers
+
     - panther_mongodb_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
