+++ packs/msft_graph.yml
@@ -9,7 +9,7 @@ PackDefinition:
     - Microsoft365.MFA.Disabled
     - Microsoft365.Exchange.External.Forwarding
     # Globals
-    - panther_base_helpers
+    - panther_msft_helpers
     - panther_config
     - panther_config_defaults
     - panther_config_overrides
