+++ packs/multisource_correlations.yml
@@ -27,8 +27,6 @@ PackDefinition:
   
   # Global Helpers
     - global_filter_github
-    - panther_base_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
+
     - panther_event_type_helpers
