+++ packs/notion.yml
@@ -24,14 +24,10 @@ PackDefinition:
     # Globals used in these detections
     - global_filter_notion
     - panther_base_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
     - panther_event_type_helpers
     - panther_ipinfo_helpers
     - panther_lookuptable_helpers
     - panther_notion_helpers
-    - panther_oss_helpers
     # Data Model
     - Standard.Notion.AuditLogs
 DisplayName: "Panther Notion Pack"
