+++ packs/okta.yml
@@ -29,11 +29,8 @@ PackDefinition:
     - Okta.Support.Reset
     # Globals used in these detections
     - panther_base_helpers
-    - panther_oss_helpers
     - panther_event_type_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+    - panther_okta_helpers
     # Data Model
     - Standard.Okta.SystemLog
 DisplayName: "Panther Okta Pack"
