+++ packs/onelogin.yml
@@ -19,10 +19,6 @@ PackDefinition:
     - OneLogin.Login
     # Globals used in these detections
     - panther_base_helpers
-    - panther_oss_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
     - panther_event_type_helpers
     # Data Model
     - Standard.OneLogin.Events
