+++ packs/onepassword.yml
@@ -9,8 +9,6 @@ PackDefinition:
     # 1Password Specific Rules
     - OnePassword.Unusual.Client
     # Supporting Global Helpers
-    - panther_base_helpers
+
     - panther_event_type_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
