+++ packs/osquery.yml
@@ -16,8 +16,6 @@ PackDefinition:
     - Osquery.SSHListener
     - Osquery.SuspiciousCron
     # Globals used in these detections
-    - panther_base_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
+
 DisplayName: "Panther OSQuery Pack"
