+++ packs/panther.yml
@@ -10,9 +10,7 @@ PackDefinition:
     # Data Model
     - Standard.Panther.Audit
     # Helpers
-    - panther_base_helpers
+
     - panther_event_type_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
 DisplayName: "Panther Audit Logs Pack"
