+++ packs/sentinelone.yml
@@ -6,8 +6,6 @@ PackDefinition:
     - SentinelOne.Alert.Passthrough
     - SentinelOne.Threats
     # Globals used in these detections
-    - panther_base_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
+
 DisplayName: "Panther SentinelOne Pack"
