+++ packs/slack.yml
@@ -27,10 +27,6 @@ PackDefinition:
     - Slack.AuditLogs.UserPrivilegeEscalation
     - Slack.AuditLogs.UserPrivilegeChangedToUser
     # Globals used in these rules/policies
-    - panther_base_helpers
-    - panther_oss_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+    - panther_slack_helpers
     # Data Model
     - Standard.Slack.AuditLogs
