+++ packs/snyk.yml
@@ -16,8 +16,6 @@ PackDefinition:
     - Snyk.User.Management
     # Globals
     - global_filter_snyk
-    - panther_base_helpers
+
     - panther_snyk_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
