+++ packs/tailscale.yml
@@ -7,10 +7,8 @@ PackDefinition:
     - Tailscale.HTTPS.Disabled
     - Tailscale.Machine.Approval.Requirements.Disabled
     # Globals used in these detections
-    - panther_base_helpers
+
     - panther_tailscale_helpers
     - global_filter_tailscale
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
 DisplayName: "Panther Tailscale Pack"
