+++ packs/tines.yml
@@ -15,8 +15,6 @@ PackDefinition:
     - Tines.Actions.DisabledChanges
     # Globals
     - global_filter_tines
-    - panther_base_helpers
+
     - panther_tines_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
