+++ packs/tor.yml
@@ -7,7 +7,4 @@ PackDefinition:
     - panther_base_helpers
     - panther_lookuptable_helpers
     - panther_tor_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
 DisplayName: "Tor Lookup Tables"
