+++ packs/wiz.yml
@@ -23,6 +23,4 @@ PackDefinition:
     - Wiz.CICD.Scan.Policy.Updated.Or.Deleted
     - panther_wiz_helpers
     - panther_base_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
