+++ packs/zendesk.yml
@@ -14,8 +14,6 @@ PackDefinition:
     # Data model
     - Standard.Zendesk.AuditLog
     # Globals
-    - panther_base_helpers
+    - panther_zendesk_helpers
     - panther_event_type_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
