+++ packs/zoom.yml
@@ -15,10 +15,7 @@ PackDefinition:
     # Data Models used in these detections
     - Standard.Zoom.Operation
     # Globals used in these detections
-    - panther_base_helpers
-    - panther_config
-    - panther_config_defaults
-    - panther_config_overrides
+
+
     - panther_event_type_helpers
-    - panther_oss_helpers
     - panther_zoom_helpers
