+++ policies/aws_rds_policies/aws_rds_instance_snapshot_public_access.py
@@ -1,4 +1,4 @@
-from panther_oss_helpers import listify
+from panther_base_helpers import listify
 
 
 def policy(resource):
